+++ tests/testthat/test-reports-run.R
@@ -29,6 +29,7 @@ test_that("run: success", {
 test_that("run: error", {
   skip_on_os("windows")
   server <- orderly_test_server()
+  on.exit(server$stop())
   remote <- montagu_server("testing", "localhost", server$port, orderly = TRUE)
   
   p <- file.path(server$path, "src", "count", "parameters.json")
@@ -54,6 +55,7 @@ test_that("run: error", {
 test_that("set timeout", {
   skip_on_os("windows")
   server <- orderly_test_server("interactive")
+  on.exit(server$stop())
   remote <- montagu_server("testing", "localhost", server$port, orderly = TRUE)
   
   p <- file.path(server$path, "src", "count", "parameters.json")